DataViews: Allow column re-ordering
